move app into folders
