+++ ibl_alignment_gui/app/controllers/app_controller.py
@@ -8,9 +8,9 @@ import numpy as np
 import pyqtgraph as pg
 from qtpy import QtCore, QtWidgets
 
-from ibl_alignment_gui.app.app_view import AlignmentGUIView
+from ibl_alignment_gui.app.controllers.shank_controller import ShankController
 from ibl_alignment_gui.app.load_worker import Worker
-from ibl_alignment_gui.app.shank_controller import ShankController
+from ibl_alignment_gui.app.views.app_view import AlignmentGUIView
 from ibl_alignment_gui.app.widgets.custom_widgets import ColorBar
 from ibl_alignment_gui.handlers.probe_handler import (
     ProbeHandlerAllenYaml,
@@ -23,7 +23,7 @@ from ibl_alignment_gui.loaders import plot_loader
 from ibl_alignment_gui.plugins.add_plugins import Plugins
 from ibl_alignment_gui.plugins.qc_dialog import display as display_qc_dialog
 from ibl_alignment_gui.plugins.upload_dialog import display as display_upload_dialog
-from ibl_alignment_gui.utils.utils import shank_loop
+from ibl_alignment_gui.utils.helpers import shank_loop
 from iblutil.util import Bunch
