+++ ibl_alignment_gui/app/controllers/shank_controller.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pyqtgraph as pg
 
-from ibl_alignment_gui.app.shank_view import ShankView
+from ibl_alignment_gui.app.views.shank_view import ShankView
 from ibl_alignment_gui.app.widgets.custom_widgets import ColorBar
 from ibl_alignment_gui.handlers.shank_handler import ShankHandler
 from iblutil.util import Bunch
