+++ ibl_alignment_gui/app/views/app_view.py
