+++ ibl_alignment_gui/app/views/shank_view.py
