+++ ibl_alignment_gui/plugins/add_plugins.py
@@ -1,5 +1,9 @@
 from typing import TYPE_CHECKING
 
+from ibl_alignment_gui.plugins.allen.additional_plots_allen import setup as setup_allen_plots
+
+# from ibl_alignment_gui.plugins.additional_plots import setup as setup_additional_plots
+from ibl_alignment_gui.plugins.allen.custom_filters import setup as setup_custom_filters
 from ibl_alignment_gui.plugins.channel_prediction import setup as setup_channel_prediction
 from ibl_alignment_gui.plugins.cluster_features import setup as setup_cluster_features
 from ibl_alignment_gui.plugins.ephys_features import setup as setup_ephys_features
@@ -8,12 +12,8 @@ from ibl_alignment_gui.plugins.qc_dialog import setup as setup_qc_dialog
 from ibl_alignment_gui.plugins.range_controller import setup as setup_control_range
 from ibl_alignment_gui.plugins.upload_dialog import setup as setup_upload_dialog
 
-# from ibl_alignment_gui.plugins.additional_plots import setup as setup_additional_plots
-from ibl_alignment_gui.plugins.allen.custom_filters import setup as setup_custom_filters
-from ibl_alignment_gui.plugins.allen.additional_plots_allen import setup as setup_allen_plots
-
 if TYPE_CHECKING:
-    from ibl_alignment_gui.app.app_controller import AlignmentGUIController
+    from ibl_alignment_gui.app.controllers.app_controller import AlignmentGUIController
 
 
 class Plugins:
