+++ ibl_alignment_gui/plugins/allen/additional_plots_allen.py
@@ -1,14 +1,13 @@
 from types import MethodType
 from typing import TYPE_CHECKING, Any
 
-
 from ibl_alignment_gui.loaders.plot_loader import skip_missing
-from ibl_alignment_gui.utils.utils import shank_loop
+from ibl_alignment_gui.utils.helpers import shank_loop
 from iblutil.util import Bunch
 
 if TYPE_CHECKING:
-    from ibl_alignment_gui.app.app_controller import AlignmentGUIController
-    from ibl_alignment_gui.app.shank_controller import ShankController
+    from ibl_alignment_gui.app.controllers.app_controller import AlignmentGUIController
+    from ibl_alignment_gui.app.controllers.shank_controller import ShankController
 
 
 PLUGIN_NAME = 'Additional Plots Allen'
