+++ ibl_alignment_gui/plugins/upload_dialog.py
@@ -5,7 +5,7 @@ from qtpy import QtWidgets
 from ibl_alignment_gui.app.widgets.custom_widgets import CheckBoxGroup
 
 if TYPE_CHECKING:
-    from ibl_alignment_gui.app.app_controller import AlignmentGUIController
+    from ibl_alignment_gui.app.controllers.app_controller import AlignmentGUIController
 
 PLUGIN_NAME = 'Upload dialog'
